feat: Enable nullboot fallback mechanism using BootNext
Per [0], we are implementing CPC0099 [1]. The BootOrder EFI variable is
now only changed if there is no new kernel or via flag activation.
Otherwise, since nullboot is called upon kernel update/upgrade, the
default behavior is to set the new kernel as the BootNext entry to
ensure a failure to boot results in a fallback to the original,
unmodified BootOrder.

Refs: [0] https://warthogs.atlassian.net/browse/CPC-8320
Refs: [1] https://docs.google.com/document/d/1SMfkBpZHC-bTYP0a9kVWlokyWsLfZy_JXfKFqh7BWTU/edit?pli=1&tab=t.0