feat: Enable nullboot fallback mechanism using BootNext
Per [0], we are implementing CPC0099 [1]. When activated, the new flag
sets BootNext as the latest versioned kernel in ESP, and does not
conduct any destructive actions in relation to the EFI boot variables or
signed assets. This is intended to be used when a new kernel is
installed to ensure the next boot is successful before removing existing
kernels or modifying the persistant BootOrder variable.

Refs: [0] https://warthogs.atlassian.net/browse/CPC-8320
Refs: [1] https://docs.google.com/document/d/1SMfkBpZHC-bTYP0a9kVWlokyWsLfZy_JXfKFqh7BWTU/edit?pli=1&tab=t.0