feat: Determine the state of the bootloader
Per [0] and [1], this change allows for a query on the state of the
bootloader. Knowing that a new kernel has been installed, but is not
installed as a boot variable on the system may indicate a circumstance
to enable the kernel fallback mechanism; likewise, knowing that
BootCurrent reflects the latest installed kernel may determine whether a
fallback occurred.

Refs: [0] https://warthogs.atlassian.net/browse/CPC-8320
Refs: [1] https://docs.google.com/document/d/1SMfkBpZHC-bTYP0a9kVWlokyWsLfZy_JXfKFqh7BWTU/edit?pli=1&tab=t.0