feat: Enable nullboot fallback mechanism using BootNext
Introduce fallback mechanism for potentially flawed kernel entries.
This method allows kernels to be booted before amending the persistent
BootOrder EFI variable. Instead of potentially adding a faulty kernel
entry to the BootOrder, BootNext can be used. If the kernel entry boots,
then the BootOrder and bootloader can be amended. If it does not boot,
then the previous BootOrder will continue to boot. After BootNext is set
and the system reboots, BootNext is automatically deleted by UEFI.