feat: CLI update to support fallback mechanism
This enables the kernel fallback mechanism and amendment method via
`nullbootctl`. The addition of these two CLI flags enables a default
kernel fallback mechanism on new kernel installations. Default behavior
is to implement the fallback mechanism, and a successful boot onto that
kernel will amend the boot loader if -confirm-kernel-boot is used.
-trust-kernel-installs essentially performs the previous nullboot
behavior.