feat: CLI update to support fallback mechanism
Per [0] and the plan laid out by [1], the addition of these two CLI
flags allows a default kernel fallback mechanism on kernel
installations. Default behavior is to implement the fallback mechanism,
and a successful boot into that kernel will ammend the boot loader
should the new start-up flag be used.

Refs: [0] https://warthogs.atlassian.net/browse/CPC-8320
Refs: [1] https://docs.google.com/document/d/1SMfkBpZHC-bTYP0a9kVWlokyWsLfZy_JXfKFqh7BWTU/edit?pli=1&tab=t.0