feat: Systemd one-shot service for Kernel fallback mechanism
Adds a systemd one-shot service to amend the bootloader on a successfully
booted, new kernel entry using the -confirm-kernel-boot option.

Should a newly installed kernel panic while booting, no changes will
occur in the boot loader.

In the event that a new kernel successfully boots, nullboot will amend
the boot loader such that BootCurrent == BootOrder[0].

However, if the current boot is not the latest kernel, nullboot will
exit with an error, but (importantly) not modify any boot variables.

In a standard reboot scenario, nullboot will run with the default
nullboot behavior, which ideally does nothing unless the user manually
modifies the EFI variables.