feat: Systemd one-shot service for Kernel fallback mechanism
Should a newly installed kernel panic while booting, no changes will
occur in the boot loader. See jira ticket [0] and specification [1].

In the event that a new kernel successfully boots, nullboot will ammend
the boot loader such that BootCurrent == BootOrder[0].

However, if the current boot is not the latest kernel, nullboot will
exit with an error, but (importantly) not modify any boot variables.

In a standard reboot scenario, nullboot will run with the default
nullboot behavior.

Refs: [0] https://warthogs.atlassian.net/browse/CPC-8320
Refs: [1] https://docs.google.com/document/d/1SMfkBpZHC-bTYP0a9kVWlokyWsLfZy_JXfKFqh7BWTU/edit?pli=1&tab=t.0